# public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dokumentasi API Penjualan",
        "description": "API untuk manajemen data penjualan",
        "contact": {
            "name": "Muhammad Gibran",
            "email": "gibran.devv@email.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/produk/{id}": {
            "get": {
                "tags": [
                    "Produk"
                ],
                "summary": "Mendapatkan detail produk berdasarkan ID",
                "operationId": "7bbd003e8c74c6e11e87515f685462a5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID produk",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mendapatkan detail produk"
                    },
                    "404": {
                        "description": "Produk tidak ditemukan"
                    }
                }
            },
            "put": {
                "tags": [
                    "Produk"
                ],
                "summary": "Memperbarui data produk",
                "operationId": "b9cce13711bd39415192f179f00718db",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID produk",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Data produk yang diperbarui",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nama_produk": {
                                        "type": "string",
                                        "example": "Laptop Gaming Terbaru"
                                    },
                                    "harga": {
                                        "type": "number",
                                        "example": "13999000"
                                    },
                                    "stok": {
                                        "type": "integer",
                                        "example": "40"
                                    },
                                    "deskripsi": {
                                        "type": "string",
                                        "example": "Laptop gaming dengan spesifikasi terkini"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Berhasil memperbarui data produk"
                    },
                    "400": {
                        "description": "Validasi gagal"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Produk"
                ],
                "summary": "Menghapus produk",
                "operationId": "ec35e29c71eb7046be687838fc1f2128",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID produk",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil menghapus produk"
                    },
                    "404": {
                        "description": "Produk tidak ditemukan"
                    }
                }
            }
        },
        "/api/produk/create": {
            "post": {
                "tags": [
                    "Produk"
                ],
                "summary": "Membuat produk baru",
                "operationId": "d0e8f1f2c6e193ec017f21bbebddbb06",
                "requestBody": {
                    "description": "Data produk baru",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nama_produk": {
                                        "type": "string",
                                        "example": "Laptop Gaming"
                                    },
                                    "harga": {
                                        "type": "number",
                                        "example": "12999000"
                                    },
                                    "stok": {
                                        "type": "integer",
                                        "example": "50"
                                    },
                                    "deskripsi": {
                                        "type": "string",
                                        "example": "Laptop gaming dengan spesifikasi tinggi"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Berhasil membuat produk baru"
                    },
                    "400": {
                        "description": "Validasi gagal"
                    }
                }
            }
        }
    },
...
}